+++ src/modules/user/dto/index.ts
@@ -2,8 +2,6 @@ export { CreateUserDto } from "./create-user.dto";
 export { LoginDto } from "./login.dto";
 export { NicknameDuplicateCheckDto } from "./nickname-duplicate-check.dto";
 export { SignUpDto } from "./sign-up.dto";
-export {
-  UserResponseDto,
-  UserTokenResponseDto,
-  NeedSignUpResponseDto,
-} from "./all-response.dto";
+export { UserResponseDto } from "./response/user-response.dto";
+export { UserTokenResponseDto } from "./response/user-token-response.dto";
+export { NeedSignUpResponseDto } from "./response/need-signup-response.dto";
